requirements.txt
flask==3.1.0
langchain==0.3.20
langchain-openai==0.3.9
python-dotenv==1.1.0
Pillow==11.1.0
openai==1.68.2
